requirements.txt
Django==5.0.6
djangorestframework==3.15.2
paypalrestsdk==1.13.3
python-dotenv==1.0.1
Requests==2.32.3
stripe==11.5.0
